fix(p2): @Transactional补全 + CRM BigDecimal化 + README统计校正
- README 修正量化指标：Java源文件 6,955（主代码）+ 365（测试），
  Controller 615，活跃模块11个，Flyway 82个，SQL文件40个，
  单元测试365个
- CRM 成交周期统计 Double→BigDecimal（CrmStatisticsCustomerDealCycleBy*RespVO）
  修复浮点精度累积误差
- HR 模块 @Transactional 补全：HrEmployeeServiceImpl/HrJobPostingServiceImpl/
  HrSocialInsuranceServiceImpl 的增删改方法添加事务保护
- MES 模块 @Transactional 补全：7 个 ServiceImpl（假期/设备/物料批次/
  质检指标/模板/工具/委外发料明细）的数据库操作方法添加事务保护
- WMS 测试修复：添加 InspectionOrderApi mock，补全测试 schema 字段
  （qc_biz_id/qc_biz_type/version）以匹配生产 DDL

+++ yudao-module-crm/src/main/java/cn/iocoder/yudao/module/crm/controller/admin/statistics/vo/customer/CrmStatisticsCustomerDealCycleByDateRespVO.java
@@ -3,6 +3,8 @@ package cn.iocoder.yudao.module.crm.controller.admin.statistics.vo.customer;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Schema(description = "管理后台 - CRM 客户成交周期分析(按日期) VO")
 @Data
 public class CrmStatisticsCustomerDealCycleByDateRespVO {
@@ -11,6 +13,6 @@ public class CrmStatisticsCustomerDealCycleByDateRespVO {
     private String time;
 
     @Schema(description = "成交周期", requiredMode = Schema.RequiredMode.REQUIRED, example = "1.0")
-    private Double customerDealCycle;
+    private BigDecimal customerDealCycle;
 
 }
